backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
pymongo[srv]==4.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
pydantic==2.5.2
pandas==2.1.4
numpy==1.26.2
tensorflow>=2.15.0
reportlab==4.0.8
python-docx==1.1.0
aiofiles==23.2.1
python-dotenv==1.0.1
opencv-python-headless==4.8.1.78
scipy==1.11.4
